src/requirements.txt
attrs==22.1.0
autopep8==1.7.0
black==22.8.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
idna==3.3
iniconfig==1.1.1
isort==5.10.1
mypy-extensions==0.4.3
numpy==1.23.2
packaging==21.3
pandas==1.4.4
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyarrow==9.0.0
pycodestyle==2.9.1
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
xlrd==2.0.1
